#553 Merge all spring-webmvc-* rules into one
