--- rule/spring-webmvc-3/README.md
@@ -1,11 +1,0 @@
-# SpecialAgent Rule for Spring Web MVC v3
-
-**Rule Name:** `spring:webmvc:3`
-
-## Compatibility
-
-```xml
-<groupId>org.springframework</groupId>
-<artifactId>spring-webmvc</artifactId>
-<version>[3.0.2.RELEASE,4.0.0.RELEASE)</version>
-```
