--- rule/spring-webmvc-3/src/main/java/io/opentracing/contrib/specialagent/rule/spring/webmvc3/SpringWebRegistryAgentRule.java
@@ -1,46 +1,0 @@
-/* Copyright 2019 The OpenTracing Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.opentracing.contrib.specialagent.rule.spring.webmvc3;
-
-import static net.bytebuddy.matcher.ElementMatchers.*;
-
-import io.opentracing.contrib.specialagent.AgentRule;
-import net.bytebuddy.agent.builder.AgentBuilder;
-import net.bytebuddy.agent.builder.AgentBuilder.Transformer;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.dynamic.DynamicType.Builder;
-import net.bytebuddy.implementation.bytecode.assign.Assigner.Typing;
-import net.bytebuddy.utility.JavaModule;
-
-public class SpringWebRegistryAgentRule extends AgentRule {
-  @Override
-  public AgentBuilder[] buildAgentUnchained(final AgentBuilder builder) {
-    return new AgentBuilder[] {builder
-      .type(named("org.springframework.web.servlet.HandlerExecutionChain"))
-      .transform(new Transformer() {
-        @Override
-        public Builder<?> transform(final Builder<?> builder, final TypeDescription typeDescription, final ClassLoader classLoader, final JavaModule module) {
-          return builder.visit(advice(typeDescription).to(SpringWebRegistryAgentRule.class).on(named("getInterceptors")));
-        }})};
-  }
-
-  @Advice.OnMethodExit
-  public static void enter(final @ClassName String className, final @Advice.Origin String origin, @Advice.Return(typing = Typing.DYNAMIC, readOnly = false) Object returned) {
-    if (isAllowed(className, origin))
-      returned = SpringWebMvcAgentIntercept.getInterceptors(returned);
-  }
-}
