--- rule/spring-webmvc-3/src/test/java/io/opentracing/contrib/specialagent/rule/spring/webmvc3/TestController.java
@@ -1,35 +1,0 @@
-/* Copyright 2019 The OpenTracing Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.opentracing.contrib.specialagent.rule.spring.webmvc3;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-import io.opentracing.util.GlobalTracer;
-
-@Controller
-public class TestController {
-  @RequestMapping(value = "/", method = RequestMethod.GET)
-  @ResponseBody
-  public String test() {
-    if (GlobalTracer.get().activeSpan() == null)
-      throw new RuntimeException("no active span");
-
-    return "test";
-  }
-}
