--- rule/spring-webmvc-3/src/test/webapp/WEB-INF/spring-web-servlet.xml
@@ -1,27 +1,0 @@
-<beans xmlns="http://www.springframework.org/schema/beans"
-	xmlns:context="http://www.springframework.org/schema/context"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xmlns:mvc="http://www.springframework.org/schema/mvc"
-	xsi:schemaLocation="
-        http://www.springframework.org/schema/beans     
-        http://www.springframework.org/schema/beans/spring-beans-3.2.xsd
-        http://www.springframework.org/schema/mvc 
-        http://www.springframework.org/schema/mvc/spring-mvc-3.2.xsd
-        http://www.springframework.org/schema/context 
-        http://www.springframework.org/schema/context/spring-context-3.2.xsd">
-
-	<context:component-scan base-package="io.opentracing.contrib.specialagent.rule.spring.webmvc3"/>
-
-<!--	<bean-->
-<!--		class="org.springframework.web.servlet.view.InternalResourceViewResolver">-->
-<!--		<property name="prefix">-->
-<!--			<value>/WEB-INF/views/jsp/</value>-->
-<!--		</property>-->
-<!--		<property name="suffix">-->
-<!--			<value>.jsp</value>-->
-<!--		</property>-->
-<!--	</bean>-->
-
-	<mvc:annotation-driven />
-	
-</beans>
