--- rule/spring-webmvc-4/README.md
@@ -1,11 +1,0 @@
-# SpecialAgent Rule for Spring Web MVC v4
-
-**Rule Name:** `spring:webmvc:4`
-
-## Compatibility
-
-```xml
-<groupId>org.springframework</groupId>
-<artifactId>spring-webmvc</artifactId>
-<version>[4.0.0.RELEASE,5.0.0.RELEASE)</version>
-```
