--- rule/spring-webmvc-4/src/test/java/io/opentracing/contrib/specialagent/rule/spring/webmvc4/TestController.java
@@ -1,33 +1,0 @@
-/* Copyright 2019 The OpenTracing Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.opentracing.contrib.specialagent.rule.spring.webmvc4;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import io.opentracing.util.GlobalTracer;
-
-@RestController
-@RequestMapping("/")
-public class TestController {
-  @RequestMapping("/sync")
-  public String test() {
-    if (GlobalTracer.get().activeSpan() == null)
-      throw new RuntimeException("no active span");
-
-    return "sync";
-  }
-}
