--- rule/spring-webmvc-4/src/test/webapp/WEB-INF/web.xml
@@ -1,59 +1,0 @@
-<!--
-
-    Copyright 2016-2019 The OpenTracing Authors
-
-    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
-    in compliance with the License. You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software distributed under the License
-    is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
-    or implied. See the License for the specific language governing permissions and limitations under
-    the License.
-
--->
-<web-app
-  xmlns="http://java.sun.com/xml/ns/javaee"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd"
-  version="3.0">
-  <display-name>OpenTracing Spring Interceptor tests</display-name>
-
-  <!-- dispatcher -->
-  <servlet>
-    <servlet-name>dispatcher</servlet-name>
-    <servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
-    <init-param>
-      <param-name>contextClass</param-name>
-      <param-value>org.springframework.web.context.support.AnnotationConfigWebApplicationContext</param-value>
-    </init-param>
-    <init-param>
-      <param-name>contextConfigLocation</param-name>
-      <param-value>io.opentracing.contrib.specialagent.rule.spring.webmvc4.SpringMVCConfiguration</param-value>
-    </init-param>
-    <load-on-startup>1</load-on-startup>
-    <async-supported>true</async-supported>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>dispatcher</servlet-name>
-    <url-pattern>/*</url-pattern>
-  </servlet-mapping>
-
-  <!-- put tracer into servlet context.. -->
-  <listener>
-    <listener-class>io.opentracing.contrib.specialagent.rule.spring.webmvc4.SpringMVCConfiguration</listener-class>
-  </listener>
-
-  <!-- tracing filter -->
-  <filter>
-    <filter-name>tracingFilter</filter-name>
-    <filter-class>io.opentracing.contrib.web.servlet.filter.TracingFilter</filter-class>
-    <async-supported>true</async-supported>
-  </filter>
-  <filter-mapping>
-    <filter-name>tracingFilter</filter-name>
-    <url-pattern>/*</url-pattern>
-  </filter-mapping>
-
-</web-app>
