--- rule/spring-webmvc-5/README.md
@@ -1,11 +1,0 @@
-# SpecialAgent Rule for Spring Web MVC v5
-
-**Rule Name:** `spring:webmvc:5`
-
-## Compatibility
-
-```xml
-<groupId>org.springframework</groupId>
-<artifactId>spring-webmvc</artifactId>
-<version>[5.0.0.RELEASE,LATEST]</version>
-```
