--- rule/spring-webmvc-5/pom.xml
@@ -1,105 +1,0 @@
-<!--
-  Copyright 2019 The OpenTracing Authors
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<project
-  xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <artifactId>rule</artifactId>
-    <groupId>io.opentracing.contrib.specialagent.rule</groupId>
-    <version>1.7.2-SNAPSHOT</version>
-  </parent>
-  <artifactId>spring-webmvc-5</artifactId>
-  <name>SpecialAgent Rule for Spring Web MVC 5</name>
-  <properties>
-    <sa.rule.name>spring:webmvc:5</sa.rule.name>
-    <min.version>5.0.0.RELEASE</min.version>
-    <passCompatibility>org.springframework:spring-webmvc:[${min.version},]</passCompatibility>
-    <failCompatibility>org.springframework:spring-webmvc:[3.0.2.RELEASE,${min.version})</failCompatibility>
-    <version.jetty>9.3.15.v20161220</version.jetty>
-  </properties>
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>io.opentracing.contrib.specialagent</groupId>
-        <artifactId>specialagent-maven-plugin</artifactId>
-        <configuration>
-          <presents>
-            <present>org.springframework.web.servlet.resource.HttpResource</present>
-          </presents>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-  <dependencies>
-    <dependency>
-      <groupId>org.springframework</groupId>
-      <artifactId>spring-webmvc</artifactId>
-      <version>${min.version}</version>
-      <optional>true</optional>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>io.opentracing.contrib</groupId>
-      <artifactId>opentracing-web-servlet-filter</artifactId>
-      <version>0.4.0</version>
-      <optional>true</optional>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.jetty</groupId>
-      <artifactId>jetty-server</artifactId>
-      <version>${version.jetty}</version>
-      <optional>true</optional>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.jetty</groupId>
-      <artifactId>jetty-servlet</artifactId>
-      <version>${version.jetty}</version>
-      <optional>true</optional>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.jetty</groupId>
-      <artifactId>jetty-webapp</artifactId>
-      <version>${version.jetty}</version>
-      <optional>true</optional>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-test</artifactId>
-      <version>2.1.4.RELEASE</version>
-      <optional>true</optional>
-      <scope>test</scope>
-    </dependency>
-
-    <!--
-      Putting this dependency in the end, because version 2.3 is not high enough for Jetty being
-      used for tests. This Dependency is only provided here for the sake of creating a
-      fingerprint.bin. This is the lowest version that is supported by java-web-servlet-filter.
-     -->
-    <dependency>
-      <groupId>javax.servlet</groupId>
-      <artifactId>servlet-api</artifactId>
-      <version>2.3</version>
-      <optional>true</optional>
-      <scope>provided</scope>
-    </dependency>
-  </dependencies>
-</project>
