--- rule/spring-webmvc-5/src/test/java/io/opentracing/contrib/specialagent/rule/spring/webmvc5/SpringMVCConfiguration.java
@@ -1,43 +1,0 @@
-/* Copyright 2019 The OpenTracing Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.opentracing.contrib.specialagent.rule.spring.webmvc5;
-
-import javax.servlet.ServletContextEvent;
-import javax.servlet.ServletContextListener;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-import org.springframework.web.servlet.config.annotation.DefaultServletHandlerConfigurer;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
-
-@EnableWebMvc
-@Configuration
-@Import({TestController.class})
-public class SpringMVCConfiguration extends WebMvcConfigurerAdapter implements ServletContextListener {
-  @Override
-  public void configureDefaultServletHandling(final DefaultServletHandlerConfigurer configurer) {
-    configurer.enable();
-  }
-
-  @Override
-  public void contextInitialized(final ServletContextEvent servletContextEvent) {
-  }
-
-  @Override
-  public void contextDestroyed(final ServletContextEvent servletContextEvent) {
-  }
-}
