+++ rule/spring-webmvc/README.md
@@ -1,0 +1,11 @@
+# SpecialAgent Rule for Spring Web MVC
+
+**Rule Name:** `spring:webmvc`
+
+## Compatibility
+
+```xml
+<groupId>org.springframework</groupId>
+<artifactId>spring-webmvc</artifactId>
+<version>[3.0.2.RELEASE,LATEST)</version>
+```
