+++ rule/spring-webmvc/src/main/java/io/opentracing/contrib/specialagent/rule/spring/webmvc/SpringWebRegistryAgentRule.java
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package io.opentracing.contrib.specialagent.rule.spring.webmvc5;
+package io.opentracing.contrib.specialagent.rule.spring.webmvc;
 
 import static net.bytebuddy.matcher.ElementMatchers.*;
 
@@ -23,13 +23,14 @@ import net.bytebuddy.agent.builder.AgentBuilder.Transformer;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.dynamic.DynamicType.Builder;
+import net.bytebuddy.implementation.bytecode.assign.Assigner.Typing;
 import net.bytebuddy.utility.JavaModule;
 
 public class SpringWebRegistryAgentRule extends AgentRule {
   @Override
   public AgentBuilder[] buildAgentUnchained(final AgentBuilder builder) {
     return new AgentBuilder[] {builder
-      .type(named("org.springframework.web.servlet.config.annotation.InterceptorRegistry"))
+      .type(named("org.springframework.web.servlet.HandlerExecutionChain"))
       .transform(new Transformer() {
         @Override
         public Builder<?> transform(final Builder<?> builder, final TypeDescription typeDescription, final ClassLoader classLoader, final JavaModule module) {
@@ -37,9 +38,9 @@ public class SpringWebRegistryAgentRule extends AgentRule {
         }})};
   }
 
-  @Advice.OnMethodEnter
-  public static void enter(final @ClassName String className, final @Advice.Origin String origin, final @Advice.This Object thiz) {
+  @Advice.OnMethodExit
+  public static void enter(final @ClassName String className, final @Advice.Origin String origin, @Advice.Return(typing = Typing.DYNAMIC, readOnly = false) Object returned) {
     if (isAllowed(className, origin))
-      SpringWebMvcAgentIntercept.getInterceptors(thiz);
+      returned = SpringWebMvcAgentIntercept.getInterceptors(returned);
   }
 }
