+++ rule/spring-webmvc/src/test/webapp/WEB-INF/spring-web-servlet.xml
@@ -1,32 +1,27 @@
-<!--
-  Copyright 2018 The OpenTracing Authors
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
 <beans xmlns="http://www.springframework.org/schema/beans"
 	xmlns:context="http://www.springframework.org/schema/context"
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xmlns:mvc="http://www.springframework.org/schema/mvc"
 	xsi:schemaLocation="
-        http://www.springframework.org/schema/beans
+        http://www.springframework.org/schema/beans     
         http://www.springframework.org/schema/beans/spring-beans-3.2.xsd
-        http://www.springframework.org/schema/mvc
+        http://www.springframework.org/schema/mvc 
         http://www.springframework.org/schema/mvc/spring-mvc-3.2.xsd
-        http://www.springframework.org/schema/context
+        http://www.springframework.org/schema/context 
         http://www.springframework.org/schema/context/spring-context-3.2.xsd">
 
 	<context:component-scan base-package="io.opentracing.contrib.specialagent.rule.spring.webmvc"/>
 
-	<mvc:annotation-driven />
+<!--	<bean-->
+<!--		class="org.springframework.web.servlet.view.InternalResourceViewResolver">-->
+<!--		<property name="prefix">-->
+<!--			<value>/WEB-INF/views/jsp/</value>-->
+<!--		</property>-->
+<!--		<property name="suffix">-->
+<!--			<value>.jsp</value>-->
+<!--		</property>-->
+<!--	</bean>-->
 
+	<mvc:annotation-driven />
+	
 </beans>
