+++ rule/spring-webmvc/src/test/webapp/WEB-INF/web.xml
@@ -1,17 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
 <!--
-  Copyright 2018 The OpenTracing Authors
 
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
+    Copyright 2016-2019 The OpenTracing Authors
 
-      http://www.apache.org/licenses/LICENSE-2.0
+    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+    in compliance with the License. You may obtain a copy of the License at
 
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
+    http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software distributed under the License
+    is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+    or implied. See the License for the specific language governing permissions and limitations under
+    the License.
+
 -->
 <web-app xmlns="http://java.sun.com/xml/ns/javaee"
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
@@ -23,11 +24,28 @@
     <servlet-name>spring-web</servlet-name>
     <servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
     <load-on-startup>1</load-on-startup>
+    <!--
+    <init-param>
+      <param-name>contextConfigLocation</param-name>
+      <param-value>/WEB-INF/spring-mvc-config.xml</param-value>
+    </init-param>
+     -->
   </servlet>
 
   <servlet-mapping>
     <servlet-name>spring-web</servlet-name>
     <url-pattern>/</url-pattern>
   </servlet-mapping>
+
+  <!-- tracing filter -->
+  <filter>
+    <filter-name>tracingFilter</filter-name>
+    <filter-class>io.opentracing.contrib.web.servlet.filter.TracingFilter</filter-class>
+    <async-supported>true</async-supported>
+  </filter>
+  <filter-mapping>
+    <filter-name>tracingFilter</filter-name>
+    <url-pattern>/*</url-pattern>
+  </filter-mapping>
 
 </web-app>
